Consolidate tracing initialization into test_helpers.rs
Move the tracing initialization function `init` to `test_helpers.rs` for improved maintainability and reuse. This change removes duplicate code across various test files and centralizes the tracing setup.
